Remove booting logic
The initial assumption was to use `.unavailable_replicas` to guess the
amount of booting replicas. This figure also contains the replicas that
are in the shutdown state.

Because the poll interval is greater than the average boot time
(seconds) and the graceful shutdown period is very high, there is a huge
chance that all of the `.unavailable_replicas` are the replicas that are
in shutdown state, not booting state.