Reduce parallelization of gecko signing and beetmover scriptworkers (#151)
We've been seeing a lot of issues with hg.m.o load lately. One of the factors here is the bursty requests that chain of trust verification makes to fetch `/json-pushes` and `.taskcluster.yml`.

This patch cuts gecko-3 beetmover & signing max_replicas in half to 40 and 50 respectively, which are by far the largest pools (next closest is 20). This will slow down nightlies and releases to some degree, but hopefully will help hg.mozilla.org stay up.