fix: set a more reasonable value for bitrisescript average_task_duration
It's impossible to know the average task duration of these workers,
because they entirely depend on how long the workflows take in Bitrise.
It could be one minute or multiple hours. But 30 minutes seems like a
reasonable middle ground.

Also bump up the SLO value so we don't spin up new nodes as quickly.