Don't scale down while there are claimed tasks (bug 1926654)
Scaling down as soon as the pending count goes to 0 has been a source of
problems over the years because it requires a long termination grace
period for containers and keeping that in sync in multiple places.
Instead, we can wait until no tasks are either pending or claimed before
we scale down.  Individual task run time will still be limited by
scriptworker's task_max_timeout, which is easier to manage.